Add files via upload
The competition hosted at GDSC IIITD acted as a source of inspiration to the project. We hope to win to bring our idea to implementation and make a very practical and import idea to this world live. In future smart contracting with IPFS and Filecoin will be done. The future update will then enable us to actually buy even real estate online without any need of paper work and physical presence of owner.